feat(#3713): add related component cards to component pages
Surface relatedComponents MDX data as mini cards below tabs on each
component page. Cards show thumbnail, name, description, and category
badge. Extract shared thumbnail utility from ComponentsGrid. Fix
relatedComponents data: typo in input, add missing bidirectional
relationships for pagination, drawer, form-item, side-menu, button.
