+++ docs/src/content/components/button.mdx
@@ -10,6 +10,7 @@ tags:
 relatedComponents:
   - button-group
   - icon-button
+  - link-button
 figmaUrl: https://www.figma.com/design/3pb2IK8s2QUqWieH79KdN7/%E2%9D%96-Component-library-%7C-DDD?node-id=420-6810
 githubUrl: https://github.com/GovAlta/ui-components/issues?q=is%3Aissue+is%3Aopen+label%3AButton
 webComponentTag: goa-button
