+++ docs/src/content/components/drawer.mdx
@@ -11,6 +11,7 @@ tags:
   - sidepanel
 relatedComponents:
   - modal
+  - push-drawer
 githubUrl: https://github.com/GovAlta/ui-components/labels/Drawer
 webComponentTag: goa-drawer
 reactClassName: GoabDrawer
