+++ docs/src/content/components/form-item.mdx
@@ -11,6 +11,11 @@ relatedComponents:
   - input
   - dropdown
   - checkbox
+  - checkbox-list
+  - date-picker
+  - file-upload-input
+  - radio-group
+  - text-area
 figmaUrl: https://www.figma.com/design/3pb2IK8s2QUqWieH79KdN7/%E2%9D%96-Component-library-%7C-DDD?node-id=27284-300347
 githubUrl: https://github.com/GovAlta/ui-components/labels/Form%20item
 webComponentTag: goa-form-item
