+++ docs/src/content/components/input.mdx
@@ -11,7 +11,7 @@ tags:
   - text input
 relatedComponents:
   - form-item
-  - textarea
+  - text-area
   - dropdown
 figmaUrl: https://www.figma.com/design/3pb2IK8s2QUqWieH79KdN7/%E2%9D%96-Component-library-%7C-DDD?node-id=27301-303447
 githubUrl: https://github.com/GovAlta/ui-components/labels/Input
@@ -25,6 +25,7 @@ Capture text input from users in forms.
 ## When to use
 
 Use Input when you need users to enter:
+
 - Short text answers (names, email addresses)
 - Numbers (phone numbers, quantities)
 - Dates and times
@@ -33,6 +34,7 @@ Use Input when you need users to enter:
 ## When not to use
 
 Don't use Input when:
+
 - Users need to enter multiple lines of text (use Textarea)
 - Users need to select from predefined options (use Dropdown or Radio)
 - The input requires complex formatting (consider specialized components)
