+++ docs/src/content/components/pagination.mdx
@@ -7,6 +7,9 @@ tags:
   - pager
   - pagination controls
   - structure and navigation
+relatedComponents:
+  - data-grid
+  - table
 figmaUrl: https://www.figma.com/design/3pb2IK8s2QUqWieH79KdN7/%E2%9D%96-Component-library-%7C-DDD?node-id=622-13964
 githubUrl: https://github.com/GovAlta/ui-components/labels/Pagination
 webComponentTag: goa-pagination
