+++ docs/src/content/components/side-menu.mdx
@@ -11,6 +11,7 @@ tags:
 relatedComponents:
   - side-menu-group
   - side-menu-heading
+  - work-side-menu
 figmaUrl: https://www.figma.com/design/3pb2IK8s2QUqWieH79KdN7/%E2%9D%96-Component-library-%7C-DDD?node-id=24089-474089
 githubUrl: https://github.com/GovAlta/ui-components/labels/Side%20menu
 webComponentTag: goa-side-menu
